Move all shared stuff to a shared folder
