+++ src/dns-tool/scss/style.scss
@@ -16,5 +16,5 @@ limitations under the License.
 
 .dns-tool {
   $header: #005ff8;
-  @import "../../scss/style.scss";
+  @import "../../shared/scss/style.scss";
 }
