+++ src/dns-tool/templates/app.vue
@@ -55,11 +55,11 @@ limitations under the License.
     import i18n from "../i18n"
     import { reports } from "../plain_text_reports"
     import RecordSelectionModal from "./record_selection_modal"
-    import cfDNS from "../../utils/cfDNS"
+    import cfDNS from "../../shared/utils/cfDNS"
     import NoSearch from "./skeletons/no_search"
     import RecordSkeleton from "./skeletons/record"
-    import Footer from "../../templates/footer"
-    import Header from "../../templates/header"
+    import Footer from "../../shared/templates/footer"
+    import Header from "../../shared/templates/header"
 
     const stripHttps = /(https*:\/\/)*(.+)*/
     const isHostname = /.*\.[a-z]+/
