+++ src/shared/i18n/en/index.ts
