+++ src/shared/i18n/en/templates/footer.ts
