+++ src/shared/i18n/en/templates/gh_link.ts
