+++ src/shared/i18n/en/templates/index.ts
