+++ src/shared/scss/_base.scss
