+++ src/shared/scss/_no-search.scss
