+++ src/shared/scss/style.scss
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 @import "globals";
-@import "../../node_modules/bulma/bulma.sass";
+@import "../../../node_modules/bulma/bulma";
 @import "base";
 @import "typography";
 @import "header";
