+++ src/shared/templates/ext_link.vue
