+++ src/shared/templates/footer.vue
