+++ src/shared/templates/gh_link.vue
