+++ src/shared/utils/dataUrlParser.ts
