+++ src/shared/utils/skeletonStyle.ts
