+++ src/spf-explainer/scss/style.scss
@@ -16,7 +16,7 @@ limitations under the License.
 
 .spf-explainer {
   $header: #0071FE;
-  @import '../../scss/style.scss';
+  @import '../../shared/scss/style.scss';
 
   .main.container {
     .record-group {
